requirements.txt
python-telegram-bot
requests
python-dotenv
rarfile
pyzbar pillow